requirements.txt
mido
pretty_midi
watchdog
pydantic
PyYAML
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
fastapi
uvicorn
faster-whisper
whisperx==3.1.1 ; platform_system!="Windows"
aeneas==1.7.3.0 ; platform_system!="Windows"
